requirements.txt
Flask==2.3.2
Flask_Bootstrap==3.3.7.1
mysql_connector_repackaged==0.3.1
